src/app/requirements.txt
streamlit
torch
torchvision
pillow==11.0.0
transformers
faiss-cpu
numpy
gdown